pyproject.toml
[project]
name = "multi_agent_code_analysis_system"
version = "0.1.0"
description = "A sophisticated multi-agent code analysis system."
requires-python = ">=3.12"

dependencies = [
    "modal",
    "fastapi",
    "httpx",
    "llama-index-callbacks-arize-phoenix",
    "arize-phoenix",
    "llama-index-llms-openai-like>=0.4.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "llama-index-program-lmformatenforcer>=0.3.0",
    "openinference-instrumentation-llama_index",
    "asyncio>=3.4.3",
    "gradio[mcp]>=5.33.0",
    "gradio-client>=1.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[tool.pytest.ini_options]
python_files = "test_*.py tests_*.py *_test.py *_tests.py"
python_classes = "Test* Tests*"
python_functions = "test*"

[tool.hatch.build.targets.wheel]
packages = [
    "src/agents",
    "src/core",
    "src/mcp_server",
    "src/orchestrator"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
